Enhanced UI; added error messages
